v5.x-scala-js: refactor Coll extensions
