+++ graph-ir/src/test/scala/special/collections/CollBenchmark.scala
@@ -4,7 +4,7 @@ import org.scalameter.{execution, Executor}
 import org.scalameter.api._
 import org.scalameter.picklers.Implicits._
 import special.collection.Coll
-import special.collection.ExtensionMethods._
+import special.collection.Extensions._
 import debox.cfor
